OCPBUGS-60903: avoid variable clash with oc-mirror
By design, `oc` replaces only the first occurence of
`_RELEASE_VERSION_LOCATION_` in the binary it's extracting. Because
`oc-mirror` v1 imports `oc` code, it ends up with multiple occurrences
of that variable:
```
❯ grep --byte-offset --only-matching --text _RELEASE_VERSION_LOCATION_ bin/oc-mirror
82752474:_RELEASE_VERSION_LOCATION_
83594512:_RELEASE_VERSION_LOCATION_
83596000:_RELEASE_VERSION_LOCATION_
146076450:_RELEASE_VERSION_LOCATION_
146715759:_RELEASE_VERSION_LOCATION_
```
During extraction of oc-mirror from the release payload, the wrong
variable will be replaced by the release version.